add stampObsolete option
Adds a stampObsolete option (defaults to off) that updates obsolete
versions with the date they became obsolete.

This option carries a performance penalty due to the additional
update required that may be significant for some use cases.

This is helpful so that you can write queries against the version
model that select between the creation and obselence date to create
a historical view of effective records as of a certain time.

Such a query will result in multiple versions if specifically
querying for the exact transition timestamp - comparing by
versionFieldTimestamp should allow you to identify the current
version.

This commit uses async and makes changes to babel to support that
usage that results in significantly different transplier output.